+++ db/migrate/20151008202843_create_collection_users.rb
@@ -1,0 +1,10 @@
+class CreateCollectionUsers < ActiveRecord::Migration
+  def change
+    create_table :collection_users do |t|
+      t.integer :user_id
+      t.integer :collection_id
+
+      t.timestamps null: false
+    end
+  end
+end