+++ CHANGELOG.md
@@ -10,6 +10,10 @@ To read more on Semantic Versioning, please see [specs here](http://semver.org/s
   - Collections appear on the user page collections tab.
   - Collections can be private or public.
   - If public, collections appear on the product's collections tab.
+    - Collection visibility can be changed via dropdown on collection box.
+  - After creation user is asked to share collection with other users.
+  - Shared, public, and owned collections visible on product/user/collection pages.
+- Add user search API endpoint for use with collection sharing. Will also be used for product sharing.
 - Add notifications system.
   - Notifications can be generated on either the frontend or backend.
     - Frontend notifications are not stored in the database and are purely alerts to tell the user that a save succeeded etc.
@@ -28,10 +32,13 @@ To read more on Semantic Versioning, please see [specs here](http://semver.org/s
 - Add correct favicons.
 - Ensure all required fields on reviews page have error messages when left blank.
 - Hitting the landing page whilst logged in automatically redirects the user to /app.
+- Field validation now occurs on new review page for fields preceeding the one clicked by the user. Errors also now appear at the end of the form.
 
 ####Bugfixes
 - Fix a number of issues that were generating console warnings related to interated components not having unique IDs.
 - Ensure welcome message is hidden when the user leaves the dashboard.
+- Add icon for collections to sidebar on product and user page.
+- Remove horizontal scrollbar from collection creation window.
 
 ### v1.0.0 [Live]